Move buildSrc/src/copied/java to buildSrc/src/java (#10087)
* Move buildSrc/src/copied/java to buildSrc/src/java

* Fix path

* Update journal abbreviation lists

* Update journal abbreviation lists

* Remove outdated journalList.mv

* fix test

---------
