+++ buildSrc/src/main/java/module-info.java
@@ -1,0 +1,3 @@
+open module org.jabref {
+    requires javafx.base;
+}
