+++ buildSrc/src/main/java/org/jabref/logic/journals/package-info.java
@@ -1,0 +1,4 @@
+/**
+ * This code is copied from src/main/java/org/jabref/logic/journals
+ */
+package org.jabref.logic.journals;
